# docs/swagger.yaml
basePath: /api/v1
host: localhost:9007
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Project that provide the data of datacenter to FeatWS
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FeatWS Resolver Bridge
  version: "1.0"
paths:
  /load:
    get:
      consumes:
      - application/json
      description: Create Resolvers description
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: Create Resolvers
      tags:
      - resolvers
  /resolve:
    post:
      consumes:
      - application/json
      description: Create Resolvers description
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: Create Resolvers
      tags:
      - resolvers
  /resolve/{resolver}:
    post:
      consumes:
      - application/json
      description: Create Resolvers description
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: Create Resolvers
      tags:
      - resolvers
  /resolvers/:
    get:
      consumes:
      - application/json
      description: Create Resolvers description
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: Create Resolvers
      tags:
      - resolvers
swagger: "2.0"
x-extension-openapi:
  example: value on a json format